+++ app/build.gradle.kts
@@ -186,4 +186,3 @@ tasks.withType<Test> {
     dependsOn("launchTestInstance")
     finalizedBy("stopTestInstance")
 }
-
