+++ app/src/main/kotlin/io/eugenethedev/taigamobile/dagger/AppComponent.kt
@@ -23,14 +23,12 @@ import io.eugenethedev.taigamobile.ui.screens.wiki.page.WikiPageViewModel
 import javax.inject.Singleton
 
 @Singleton
-@Component(modules = [DataModule::class, RepositoriesModule::class])
+@Component(modules = [DataModule::class, RepositoryModule::class])
 interface AppComponent {
 
-    @Component.Builder
-    interface Builder {
-        @BindsInstance
-        fun context(context: Context): Builder
-        fun build(): AppComponent
+    @Component.Factory
+    interface Factory {
+        fun create(@BindsInstance context: Context) : AppComponent
     }
 
     fun inject(mainViewModel: MainViewModel)
