+++ app/src/main/kotlin/io/eugenethedev/taigamobile/dagger/RepositoryModule.kt
@@ -1,0 +1,44 @@
+package io.eugenethedev.taigamobile.dagger
+
+import dagger.Binds
+import dagger.Module
+import io.eugenethedev.taigamobile.data.repositories.AuthRepository
+import io.eugenethedev.taigamobile.data.repositories.ProjectsRepository
+import io.eugenethedev.taigamobile.data.repositories.SprintsRepository
+import io.eugenethedev.taigamobile.data.repositories.TasksRepository
+import io.eugenethedev.taigamobile.data.repositories.UsersRepository
+import io.eugenethedev.taigamobile.data.repositories.WikiRepository
+import io.eugenethedev.taigamobile.domain.repositories.IAuthRepository
+import io.eugenethedev.taigamobile.domain.repositories.IProjectsRepository
+import io.eugenethedev.taigamobile.domain.repositories.ISprintsRepository
+import io.eugenethedev.taigamobile.domain.repositories.ITasksRepository
+import io.eugenethedev.taigamobile.domain.repositories.IUsersRepository
+import io.eugenethedev.taigamobile.domain.repositories.IWikiRepository
+import javax.inject.Singleton
+
+@Module
+abstract class RepositoryModule {
+    @Singleton
+    @Binds
+    abstract fun bindIAuthRepository(authRepository: AuthRepository): IAuthRepository
+
+    @Singleton
+    @Binds
+    abstract fun bindIProjectsRepository(searchRepository: ProjectsRepository): IProjectsRepository
+
+    @Singleton
+    @Binds
+    abstract fun bindIStoriesRepository(storiesRepository: TasksRepository): ITasksRepository
+
+    @Singleton
+    @Binds
+    abstract fun bindIUsersRepository(usersRepository: UsersRepository): IUsersRepository
+
+    @Singleton
+    @Binds
+    abstract fun bindISprintsRepository(sprintsRepository: SprintsRepository): ISprintsRepository
+
+    @Singleton
+    @Binds
+    abstract fun bindIWikiRepository(wikiRepository: WikiRepository): IWikiRepository
+}
